# go/cmd/policy-enforcer/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DYNAMOS Policy Enforcer API
  description: |
    API for policy enforcement in the DYNAMOS system.
    
    This API provides two levels of abstraction:
    
    ## Policy Enforcer API (`/policy-enforcer/*`)
    High-level, reasoner-agnostic interface for policy validation. This API is designed
    to work with different reasoning engines (eFLINT, Symboleo, JSON-based, etc.) through
    a common interface. Use these endpoints for:
    - **Allowed Clauses**: Query what request types, datasets, archetypes, and compute providers
      are allowed for a specific requester at an organization
    - **Request Validation**: Check if a specific data request is permitted by the policy
    
    ## eFLINT API (`/eflint/*`)
    Low-level interface for direct eFLINT server management. Use these endpoints for:
    - **Instance Management**: Start, stop, and monitor eFLINT server instances
    - **Commands**: Send raw commands to the eFLINT server
    - **State Management (POC)**: Export, import, and checkpoint eFLINT execution graphs
    
    ## Architecture
    ```
    ┌─────────────────────────────────────────────────────────────────────────┐
    │                           HTTP/gRPC Layer                                │
    ├─────────────────────────────────────────────────────────────────────────┤
    │  Policy Enforcer API              │  eFLINT API                         │
    │  /policy-enforcer/*               │  /eflint/*                          │
...
servers:
...
paths:
...
  /policy-enforcer/allowed-archetypes:
    get:
      summary: Get allowed archetypes
      description: |
        Returns all archetypes that are allowed for a requester at an organization.
        This is a reasoner-agnostic endpoint that works with any configured reasoning engine.
      operationId: getAllowedArchetypes
      tags:
        - Policy Enforcer
      parameters:
        - $ref: '#/components/parameters/OrganizationParam'
        - $ref: '#/components/parameters/RequesterParam'
      responses:
        '200':
          description: Allowed archetypes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllowedClausesResponse'
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Reasoner is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
...
  parameters:
    OrganizationParam:
      name: organization
      in: query
      required: true
      description: The organization/steward identifier
      schema:
        type: string
      example: VU
    RequesterParam:
      name: requester
      in: query
      required: true
      description: The requester/user identifier (typically an email)
      schema:
        type: string
      example: jorrit.stutterheim@cloudnation.nl
...
  schemas:
...
    AllowedClausesResponse:
      type: object
      properties:
        organization:
          type: string
          description: The organization/steward identifier
          example: VU
        requester:
          type: string
          description: The requester/user identifier
          example: jorrit.stutterheim@cloudnation.nl
        values:
          type: array
          items:
            type: string
          description: List of allowed values
          example: ["sqlDataRequest", "genericRequest"]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "no instance running"
...